+++ CHANGELOG.md
@@ -3,3 +3,6 @@
 - Rename "Description" -> "Service" #9
 - Fix elasticsearch index name #7
 - Add optional "Hostname" and "Port" elasticsearch fields #10
+- Refactoring (producer.go deleted, some minor fixes)
+- Expand env all field in Torchfile
+- Elasticsearch index viewer cli tool #2